requirements.txt
aiofiles==0.8.0
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.17.0
anyio==4.4.0
async-timeout==4.0.3
asyncpraw==7.7.1
asyncprawcore==2.4.0
attrs==23.2.0
boto3==1.34.144
botocore==1.34.144
certifi==2024.7.4
charset-normalizer==3.3.2
exceptiongroup==1.2.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jmespath==1.0.1
multidict==6.0.5
praw==7.7.1
prawcore==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.4
requests==2.32.3
s3transfer==0.10.2
schedule==1.2.2
six==1.16.0
sniffio==1.3.1
typing_extensions==4.12.2
update-checker==0.18.0
urllib3==2.2.2
websocket-client==1.8.0
yarl==1.9.4
